fix yurii's fix

+++ CalculatorBasedOnStack/StackCalculator.cs
@@ -15,7 +15,7 @@ namespace CalculatorBasedOnStack
             var elemntsFromPostFixForm = postFixForm.Split(' ');
             for (int i = 0; i < elemntsFromPostFixForm.Length; i++)
             {
-                var isNumber = double.TryParse(elemntsFromPostFixForm[i], out int number);
+                var isNumber = double.TryParse(elemntsFromPostFixForm[i], out double number);
                 if (isNumber)
                 {
                     stack.Push(number);
